refactor: decompose backend services and routes for maintainability
