+++ backend/src/routes/medication-enrichment.ts
@@ -8,7 +8,7 @@ import {
 	type MedicationEnrichmentEnrichRequest,
 	MedicationEnrichmentServiceError,
 	searchMedicationEnrichment,
-} from "../services/medication-enrichment.js";
+} from "../services/medication-enrichment/index.js";
 import {
 	applyOpenApiRouteStandards,
 	genericErrorSchema,
